+++ impl/src/test/groovy/io/jsonwebtoken/impl/DefaultJwtParserBuilderTest.groovy
@@ -357,6 +357,17 @@ class DefaultJwtParserBuilderTest {
         }
     }
 
+    @Test
+    void testSetSigningKeyWithPrivateKey() {
+        try {
+            builder.setSigningKey(TestKeys.RS256.pair.private)
+            fail()
+        } catch (UnsupportedKeyException e) {
+            String msg = 'JWS verification key must be either a SecretKey (for MAC algorithms) or a PublicKey (for Signature algorithms).'
+            assertEquals msg, e.getMessage()
+        }
+    }
+
     static class TestCompressionCodec implements CompressionCodec {
 
         String id