+++ impl/src/test/groovy/io/jsonwebtoken/impl/security/StandardSecureDigestAlgorithmsTest.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2022 jsonwebtoken.io
+ * Copyright © 2023 jsonwebtoken.io
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,23 +15,15 @@
  */
 package io.jsonwebtoken.impl.security
 
-import java.security.interfaces.RSAKey
+import org.junit.Test
 
-class TestRSAKey extends TestKey implements RSAKey {
+import static org.junit.Assert.assertNull
 
-    final def src
-    BigInteger modulus
+class StandardSecureDigestAlgorithmsTest {
 
-    TestRSAKey(def key) {
-        this.src = key
-        this.algorithm = key?.getAlgorithm()
-        this.format = key?.getFormat()
-        this.encoded = key?.getEncoded()
-        this.modulus = key?.getModulus()
-    }
-
-    @Override
-    BigInteger getModulus() {
-        return this.modulus
+    @Test
+    void testFindByPublicSigningKey() {
+        //public keys are not supported for signing:
+        assertNull StandardSecureDigestAlgorithms.findBySigningKey(new TestPublicKey())
     }
 }